Separate keybinds dialog by category
